added last updated feature using moment.js library
wanted to explore using libraries! chose Moment.js 2.29.1 because it was one of the first that showed up in the results even though more enhanced and faster performing libraries like Luxon are available. (Exploratory coding)

I added the geolocation because I got annoyed looking at an empty landing page.. it was quick!